% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_position__mae_centre_mean.R
\name{abs_error_centre_mean}
\alias{abs_error_centre_mean}
\title{Compute Mean Absolute Error Centre}
\usage{
abs_error_centre_mean(borehole)
}
\arguments{
\item{borehole}{A data.table or data frame containing the columns \code{prediction}, \code{target_variable}, \code{depth}, and \code{original_index}.}
}
\value{
A numeric vector containing the absolute distance between the predicted and target depths.
If no valid data is found, it returns \code{NA}.
}
\description{
This function calculates the absolute error in depth between the predicted and target centre of stratigraphic units
The centre of a unit is defined as the mean value of all the depths of the unit
}
