% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_position__mae_top.R
\name{abs_error_top}
\alias{abs_error_top}
\title{Compute Mean Absolute Error Top}
\usage{
abs_error_top(borehole)
}
\arguments{
\item{borehole}{A data.table or data frame containing the columns \code{prediction}, \code{target_variable}, \code{depth}, and \code{original_index}.}
}
\value{
A numeric vector containing the absolute distance between the predicted and target depths.
If no valid data is found, it returns \code{NA}.
}
\description{
This function calculates the absolute error in depth between the predicted and target top of stratigraphic units
It handles exclusions of specific target formations (e.g., "NN")
}
