% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_wrappers.R
\name{adhoc_lamina}
\alias{adhoc_lamina}
\title{Run a lamina adhoc by itself}
\usage{
adhoc_lamina(lamina_path, silent = FALSE)
}
\arguments{
\item{lamina_path}{Path to lamina}

\item{silent}{Suppress log output}
}
\value{
invisible data frame of execution plan
}
\description{
Run a lamina adhoc by itself
}
\examples{
\dontrun{
adhoc_lamina("PATH/TO/LAMINA/FOLDER/")
}
}
