% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_wrappers.R
\name{adhoc_stratum}
\alias{adhoc_stratum}
\title{Run a stratum adhoc by itself}
\usage{
adhoc_stratum(stratum_path, silent = FALSE)
}
\arguments{
\item{stratum_path}{Path to stratum}

\item{silent}{Suppress log output}
}
\value{
invisible data frame of execution plan
}
\description{
Run a stratum adhoc by itself
}
\examples{
\dontrun{
adhoc_stratum("PATH/TO/STRATUM/FOLDER/")
}
}
