% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics__helper_majority_vote.R
\name{apply_majority_vote}
\alias{apply_majority_vote}
\title{Apply Majority Vote to Rolling Window of Predictions}
\usage{
apply_majority_vote(df, window_size)
}
\arguments{
\item{df}{A data frame containing a column \code{prediction} with predicted values, and a grouping variable \code{nr} (e.g., borehole ID).}

\item{window_size}{An integer specifying the size of the rolling window over which the majority vote is applied.}
}
\value{
A data frame with the same structure as the input \code{df}, where the \code{prediction} column has been updated with the majority vote within each rolling window.
The \code{prediction} values are centered within the window.
}
\description{
This function applies a majority vote to the \code{prediction} column of a data frame (\code{df})
within a rolling window of specified size (\code{window_size}) for each borehole (\code{nr}).
The majority vote is determined using the \code{majority_vote} function,
which returns the most frequent value within each window.
The function is applied to each group defined by \code{nr} (borehole).
}
