% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_build.R
\name{build_outlined_strata_project}
\alias{build_outlined_strata_project}
\title{Build a strata project from an outline}
\usage{
build_outlined_strata_project(outline)
}
\arguments{
\item{outline}{A data frame with the following columns: project_path,
stratum_name, stratum_order, lamina_name, lamina_order, skip_if_fail}
}
\value{
invisible dataframe of the survey of the strata project
}
\description{
Build a strata project from an outline
}
\examples{
outline <- tibble::tibble(
  project_path = fs::dir_create(fs::file_temp()),
  stratum_name = "test",
  stratum_order = 1,
  lamina_name = "test",
  lamina_order = 1,
  skip_if_fail = FALSE
)
result <- build_outlined_strata_project(outline)
dplyr::glimpse(result)
}
