% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pipeline.R
\name{build_pipeline}
\alias{build_pipeline}
\title{Add a pipeline skeleton to your project space}
\usage{
build_pipeline(pipeline_name, path = ".")
}
\arguments{
\item{pipeline_name}{A string that is the name of your pipeline}

\item{path}{a path to where you want to drop your pipeline}
}
\value{
invisibly returns fs::path to pipeline
}
\description{
Add a pipeline skeleton to your project space
}
\examples{
build_pipeline("my_pipeline", "PATH/TO/PROJECT/FOLDER/")
}
