% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_build.R
\name{build_quick_strata_project}
\alias{build_quick_strata_project}
\title{Quickly build strata project with minimal input and standard names}
\usage{
build_quick_strata_project(project_path, num_strata = 1, num_laminae_per = 1)
}
\arguments{
\item{project_path}{Path to the project folder}

\item{num_strata}{Number of strata to create}

\item{num_laminae_per}{Number of laminae to create per stratum}
}
\value{
invisible dataframe of the survey of the strata project
}
\description{
Quickly build strata project with minimal input and standard names
}
\examples{
tmp <- fs::dir_create(fs::file_temp())
result <- build_quick_strata_project(tmp, 2, 2)
dplyr::glimpse(result)
}
