% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{build_stratum}
\alias{build_stratum}
\title{Add a stratum skeleton to your project space}
\usage{
build_stratum(stratum_name, path = ".", order = 1)
}
\arguments{
\item{stratum_name}{A string that is the name of your stratum}

\item{path}{a path to where you want to drop your stratum}

\item{order}{the order of the stratum}
}
\value{
invisibly returns fs::path to stratum
}
\description{
Add a stratum skeleton to your project space
}
\examples{
build_stratum("my_stratum_name", "PATH/TO/PROJECT/FOLDER/")
}
