% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_traditional__w_f1.R
\name{calculate_w_f1}
\alias{calculate_w_f1}
\title{Calculate Weighted F1 Score}
\usage{
calculate_w_f1(df)
}
\arguments{
\item{df}{A data frame containing \code{target_variable} (true labels) and \code{prediction} (predicted labels).}
}
\value{
The overall weighted F1 score.
}
\description{
This function calculates the weighted F1 score for each class in a classification task
and returns the overall weighted F1 score.
}
