% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_traditional__macro_f1_score.R
\name{compute_f1_score}
\alias{compute_f1_score}
\title{Compute Macro F1 Score}
\usage{
compute_f1_score(predictions, true_labels)
}
\arguments{
\item{predictions}{A vector of predicted class labels.}

\item{true_labels}{A vector of true class labels.}
}
\value{
The macro-averaged F1 score.
}
\description{
This function calculates the macro-averaged F1 score for a classification task.
}
