% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing__data.R
\name{copy_next_previous_row_info}
\alias{copy_next_previous_row_info}
\title{Copy information from the next and previous rows for multiple columns}
\usage{
copy_next_previous_row_info(df, cols)
}
\arguments{
\item{df}{Merged dataframe with basisdata}

\item{cols}{Vector of column names to copy information from the next and previous rows}
}
\value{
Dataframe with copied information from the next and previous rows
}
\description{
This function calculates the index of the next and previous rows within each borehole group,
copies the information from the next and previous rows for the specified columns, and renames
the columns accordingly.
}
