% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sequence__transition_validation_score.R
\name{correct_transitions}
\alias{correct_transitions}
\title{Compare Extracted Transitions to Known Transitions}
\usage{
correct_transitions(x, known_transitions)
}
\arguments{
\item{x}{A character vector of extracted transitions. Each transition should be a string representing
the concatenation of adjacent elements.}

\item{known_transitions}{A character vector of known transitions. This vector contains the transitions
that are considered correct or valid.}
}
\value{
A numeric value representing the proportion of transitions in \code{x} that are also present in
\code{known_transitions}.
}
\description{
This function compares a vector of extracted transitions to a set of known transitions and calculates
the proportion of correctly matched transitions. The function returns the proportion of transitions
that are found in the set of known transitions.
}
