% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction__neural_network.R
\name{custom_accuracy}
\alias{custom_accuracy}
\title{Custom Accuracy Metric for Classification}
\usage{
custom_accuracy(y_true, y_pred)
}
\arguments{
\item{y_true}{A TensorFlow tensor containing the true labels in one-hot encoded format.}

\item{y_pred}{A TensorFlow tensor containing the predicted probabilities.}
}
\value{
A TensorFlow tensor representing the mean accuracy, where values range from 0 to 1.
}
\description{
This function computes the accuracy of a classification model by comparing the predicted and true labels.
It uses the argmax of the predicted and true labels to determine the class with the highest probability and checks if they match.
The accuracy is then calculated as the mean of the correct predictions.
}
