% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing__data.R
\name{dgm_to_nap}
\alias{dgm_to_nap}
\title{Adjust TOP and BASIS to NAP (Normal Amsterdams Peil)}
\usage{
dgm_to_nap(dgm_extractie)
}
\arguments{
\item{dgm_extractie}{geodataframe DGM_extractie}
}
\value{
A GeoDataFrame
}
\description{
Adjust TOP and BASIS to NAP (Normal Amsterdams Peil)
}
