% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing__data.R
\name{discretize_intervals}
\alias{discretize_intervals}
\title{Discretize Intervals}
\usage{
discretize_intervals(basis_dgm_roer_combined, interval = 0.5)
}
\arguments{
\item{basis_dgm_roer_combined}{A data frame containing the original data with columns \code{nr}, \code{top}, and \code{bottom}.}

\item{interval}{A numeric value specifying the interval for sequence generation. Default is 0.5.}
}
\value{
A data.table with discretized sample points matched to the original data.
}
\description{
This function discretizes intervals for each unique \code{nr} in the \code{basis_dgm_roer_combined} dataset.
It calculates the highest \code{top} and lowest \code{bottom} values for each \code{nr}, creates sequences
between these values at specified intervals, and matches the sequences with the original data.
}
