% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_toml_interaction.R
\name{edit_toml}
\alias{edit_toml}
\title{Edit a toml file by providing a dataframe of desired contents}
\usage{
edit_toml(original_toml_path, new_toml_dataframe)
}
\arguments{
\item{original_toml_path}{Path to the original toml file}

\item{new_toml_dataframe}{Dataframe of the new toml file contents}
}
\value{
invisible original toml file path to toml file
}
\description{
Edit a toml file by providing a dataframe of desired contents
}
\examples{
\dontrun{
edit_toml("PATH/TO/ORIGINAL/TOML/FILE.toml", new_toml_dataframe)
}
}
