% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sequence__transition_validation_score.R
\name{extract_transitions}
\alias{extract_transitions}
\title{Extract Transitions Between Elements in a Vector}
\usage{
extract_transitions(x)
}
\arguments{
\item{x}{A character vector. This vector should contain the elements between which transitions
are to be extracted.}
}
\value{
A list containing a single character vector of transitions. Each transition is a string
formed by concatenating adjacent elements in the input vector.
}
\description{
This function generates a list of transitions between consecutive elements in a character vector.
Each transition is represented as a concatenation of adjacent elements in the original vector.
}
