% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-core.R
\name{format_timestamp}
\alias{format_timestamp}
\title{Format timestamp for log messages}
\usage{
format_timestamp(time)
}
\arguments{
\item{time}{A POSIXct time object}
}
\value{
A formatted timestamp string
}
\description{
Format timestamp for log messages
}
\keyword{internal}
