% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing__strat_colors.R
\name{get_strat_colors}
\alias{get_strat_colors}
\title{Get Stratigraphic Unit Colors}
\usage{
get_strat_colors(path_to_colors, stratigraphic_order)
}
\arguments{
\item{path_to_colors}{A string specifying the path to the CSV file containing color information.}

\item{stratigraphic_order}{A character vector defining the desired order of stratigraphic units.}
}
\value{
A named vector where stratigraphic unit names are mapped to their corresponding RGB colors.
}
\description{
This function reads a color mapping file and returns a named color vector for stratigraphic units.
}
