% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_traditional__cohens_kappa.R
\name{kappa_metric}
\alias{kappa_metric}
\title{Compute Cohen's Kappa Statistic}
\usage{
kappa_metric(df)
}
\arguments{
\item{df}{A data frame containing \code{target_variable} (true labels) and \code{prediction} (predicted labels).}
}
\value{
The Cohen's Kappa statistic.
}
\description{
This function calculates Cohen's Kappa statistic to measure inter-rater agreement.
}
