% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing__load_datasets.R
\name{load_dataset}
\alias{load_dataset}
\title{Load basisdata and transform into a GeoDataFrame}
\usage{
load_dataset(
  file_path,
  crs_target,
  cords_names = c("x", "y"),
  area_of_interest
)
}
\arguments{
\item{file_path}{path to basisdata}

\item{crs_target}{CRS code}

\item{cords_names}{list with names of x and y parameter}
}
\value{
A GeoDataFrame
}
\description{
Load basisdata and transform into a GeoDataFrame
}
