% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_error.R
\name{log_error}
\alias{log_error}
\title{Wrapper around log_message for ERROR messages in the log}
\usage{
log_error(message)
}
\arguments{
\item{message}{A string to print to stderr}
}
\value{
A message is printed to stdout or stderr
}
\description{
Wrapper around log_message for ERROR messages in the log
}
\examples{
log_message("This is an error message")
}
