% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_message.R
\name{log_message}
\alias{log_message}
\title{Send a standardized message to stdout or stderr}
\usage{
log_message(message, level = "INFO", out_or_err = "OUT")
}
\arguments{
\item{message}{A string message to log}

\item{level}{The level of the message (e.g. INFO, WARNING, ERROR)}

\item{out_or_err}{Whether to log to stdout or stderr}
}
\value{
A message is printed to stdout or stderr
}
\description{
Send a standardized message to stdout or stderr
}
\examples{
log_message("This is an info message", "INFO", "OUT")
}
