% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_total_time}
\alias{log_total_time}
\title{Print time difference in a standard for the log}
\usage{
log_total_time(begin, end)
}
\arguments{
\item{begin}{A data-time object}

\item{end}{A data-time object}
}
\value{
A numeric value of the time difference in seconds
}
\description{
Print time difference in a standard for the log
}
\examples{
log_total_time(Sys.time(), Sys.time() + 999)
}
