% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Entry point into your strata project and automation target}
\usage{
main(project_path = NULL)
}
\arguments{
\item{project_path}{A path to automation project folder}
}
\value{
invisible execution plan
}
\description{
Entry point into your strata project and automation target
}
\examples{
\dontrun{
main("/PATH/TO/PROJECT/FOLDER")
}
}
