% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics__helper_majority_vote.R
\name{majority_vote}
\alias{majority_vote}
\title{Majority Vote for Classification}
\usage{
majority_vote(x)
}
\arguments{
\item{x}{A vector of values (e.g., class labels) for which the majority vote needs to be determined.}
}
\value{
A single value representing the most frequent value (class) in the input vector \code{x}.
}
\description{
This function returns the most frequent value (or class) in a vector \code{x}. If there are ties, it returns the first occurring value
that has the maximum count.
}
