% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_traditional__accuracy.R
\name{mean_per_borehole_accuracy_custom_metric}
\alias{mean_per_borehole_accuracy_custom_metric}
\title{Compute per borehole Accuracy Metric}
\usage{
mean_per_borehole_accuracy_custom_metric(df)
}
\arguments{
\item{df}{A data frame containing \code{target_variable} (true labels) and \code{prediction} (predicted labels).}
}
\value{
The mean per borehole accuracy.
}
\description{
This function calculates the mean per borehole accuracy of a classification model while ignoring rows where \code{target_variable} is "NN".
A warning is issued if any rows are removed during filtering.
}
