% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing__data.R
\name{merge_basis_dgm}
\alias{merge_basis_dgm}
\title{Assigns "strat_inter" values to basisdata based on dgm_extractie (Normal Amsterdams Peil)}
\usage{
merge_basis_dgm(basisdata, dgm_extractie, stratigraphic_order)
}
\arguments{
\item{basisdata}{dataframe with basisdata}

\item{dgm_extraction}{dataframe with dgm_extractie}
}
\value{
basis_dgm_roer_combined
}
\description{
Assigns "strat_inter" values to basisdata based on dgm_extractie (Normal Amsterdams Peil)
}
