% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics__evaluation.R
\name{metrics_evaluation}
\alias{metrics_evaluation}
\title{Evaluate Various Metrics for Borehole Predictions}
\usage{
metrics_evaluation(df)
}
\arguments{
\item{df}{A data frame containing the \code{prediction}, \code{target_variable}, and \code{nr} (borehole ID) columns. The data is used to compute a variety of metrics.}
}
\value{
A named vector of computed metrics for each window size (original, window_3, window_5, window_7, window_9).
These include overall accuracy, F1-score, Unit match F1-score, transition match F1-score,
and various other metrics.
}
\description{
This function computes several performance metrics to evaluate the prediction accuracy for boreholes (\code{df}).
It applies a sliding window approach with various window sizes (3, 5, 7, 9) and calculates multiple metrics, including traditional and geology-inspired metrics.
It uses the \code{apply_majority_vote} function for resampling and evaluates the metrics using custom functions for different summary statistics.
}
