% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing__data.R
\name{one_hot_encoding}
\alias{one_hot_encoding}
\title{One-Hot Encoding}
\usage{
one_hot_encoding(data)
}
\arguments{
\item{data}{A data frame containing the dataset to be one-hot encoded.}
}
\value{
A data frame with one-hot encoded variables.
}
\description{
Function to perform one-hot encoding on a data frame.
}
\details{
This function converts categorical variables into a one-hot encoded format,
creating dummy variables for each category.
}
\examples{
data <- data.frame(
  nr = c(1, 2, 1, 3),
  category = c("A", "B", "A", "C")
)
one_hot_encoding(data)
}
