% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_traditional__accuracy.R
\name{overall_accuracy_custom_metric}
\alias{overall_accuracy_custom_metric}
\title{Compute Overall Accuracy Metric}
\usage{
overall_accuracy_custom_metric(df)
}
\arguments{
\item{df}{A data frame containing \code{target_variable} (true labels) and \code{prediction} (predicted labels).}
}
\value{
The overall accuracy as a proportion of correctly classified instances.
}
\description{
This function calculates the overall accuracy of a classification model while ignoring rows where \code{target_variable} is "NN".
A warning is issued if any rows are removed during filtering.
}
