% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_traditional__macro_f1_score.R
\name{overall_f1_score_custom_summary}
\alias{overall_f1_score_custom_summary}
\title{Compute Overall F1 Score Summary}
\usage{
overall_f1_score_custom_summary(df)
}
\arguments{
\item{df}{A data frame containing \code{target_variable} (true labels) and \code{prediction} (predicted labels).}
}
\value{
The macro-averaged F1 score after filtering.
}
\description{
This function computes the macro-averaged F1 score after filtering out rows where \code{target_variable} is "NN".
}
