% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_traditional__macro_f1_score.R
\name{per_borehole_f1_score_custom_summary}
\alias{per_borehole_f1_score_custom_summary}
\title{Compute per borehole F1 Score Summary}
\usage{
per_borehole_f1_score_custom_summary(df)
}
\arguments{
\item{df}{A data frame containing \code{target_variable} (true labels) and \code{prediction} (predicted labels).}
}
\value{
The mean F1 score per borehole after filtering out "NN" values.
}
\description{
This function computes the macro-averaged F1 score per borehole
}
