% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing__plot_class_proportion.R
\name{plot_class_proportion}
\alias{plot_class_proportion}
\title{Plot Class Proportion}
\usage{
plot_class_proportion(df_labeled, color_vector, stratigraphic_order)
}
\arguments{
\item{df_labeled}{A data frame containing labeled data, including the \code{strat_inter} column representing lithostratigraphic units.}

\item{color_vector}{A named vector mapping lithostratigraphic units to specific colors.}

\item{stratigraphic_order}{A character vector defining the desired order of stratigraphic units.}
}
\value{
A \code{ggplot2} object displaying class proportions.
}
\description{
This function creates a bar plot showing the proportion of each lithostratigraphic unit in the dataset.
}
