% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction__tensorflow_helper.R
\name{predictions_to_df}
\alias{predictions_to_df}
\title{Convert Tensorflow Model Predictions to DataFrame with Predicted Labels}
\usage{
predictions_to_df(predictions, df_pred, col_names)
}
\arguments{
\item{predictions}{A TensorFlow tensor containing model predictions, with shape \verb{[batch_size, num_classes]}.}

\item{df_pred}{A data frame containing the original data that the predictions correspond to.}

\item{col_names}{A character vector of column names representing the class labels corresponding to the predictions.}
}
\value{
A data frame with the original input data (\code{df_pred}) and an additional column \code{prediction},
which contains the predicted class labels based on the highest prediction probability.
}
\description{
This function converts model predictions (in tensor format) into a data frame and associates the predicted labels with the highest probability
from the predictions.
}
