% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction__neural_network.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess Data for TensorFlow Model}
\usage{
preprocess_data(df, splits_ind)
}
\arguments{
\item{df}{A data frame containing the input data and target variables. Target variables must start with \code{strat_inter_}.}

\item{splits_ind}{An integer vector indicating the row indices used for splitting the data.}
}
\value{
A list containing the following elements:
\item{value_ragged_tensor}{A TensorFlow RaggedTensor of the input feature data.}
\item{target_ragged_tensor}{A TensorFlow RaggedTensor of the target variable data.}
\item{values_names}{A character vector of column names for the input features.}
\item{target_names}{A character vector of column names for the target variables}
\item{df}{The original input data frame.}
}
\description{
This function preprocesses a data frame by separating the input features and target variables, converting them into TensorFlow constants,
and then creating ragged tensors for use in TensorFlow models. The function also prepares the necessary column names and splits the data
based on provided indices.
}
