% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics__helper_RF.R
\name{rf_metrics_by_hyper_parameter}
\alias{rf_metrics_by_hyper_parameter}
\title{Summarize and Reshape Random Forest Metrics by Hyperparameter}
\usage{
rf_metrics_by_hyper_parameter(metrics_by_fold)
}
\arguments{
\item{metrics_by_fold}{A data frame containing random forest performance metrics, with columns for \code{mtry}, \code{fold}, and various metrics (e.g., accuracy, error)
calculated during cross-validation, along with the standard deviation for each metric.}
}
\value{
A data frame with reshaped and summarized metrics, including columns for \code{mtry}, \code{metric_name}, \code{window}, \code{mean}, and \code{SD}.
The data is in long format with separate columns for the metric value and standard deviation.
}
\description{
This function processes a data frame of random forest performance metrics (\code{metrics_by_fold}) by summarizing the mean and standard deviation
of metrics for different values of the hyperparameter \code{mtry}. It then reshapes the data to a long format and separates the metric, window,
and statistic for further analysis.
}
