% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sequence__sequence_alignment.R
\name{sequence_alignment_custom_summary}
\alias{sequence_alignment_custom_summary}
\title{Sequence Alignment Score Custom Summary}
\usage{
sequence_alignment_custom_summary(df)
}
\arguments{
\item{df}{A data frame containing at least two columns: \code{prediction} (predicted sequence)
and \code{target_variable} (ground truth sequence), and \code{nr} (borehole identifier).}
}
\value{
A numeric value representing the mean SAS score for the entire dataset.
}
\description{
Computes the mean Sequence Alignment Score (SAS) for each borehole in the dataset.
The Sequence Alignment Score is calculated using the \code{sequence_alignment_score} function,
and the mean score is returned for each borehole.
}
