% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sequence__sequence_alignment.R
\name{sequence_alignment_score}
\alias{sequence_alignment_score}
\title{Sequence Alignment Score (SAS)}
\usage{
sequence_alignment_score(prediction, target)
}
\arguments{
\item{prediction}{A character vector representing the predicted sequence of categories.}

\item{target}{A character vector representing the target sequence of categories.}
}
\value{
A numeric value representing the alignment score, ranging from 0 to 1.
}
\description{
Calculates the Sequence Alignment Score (SAS) between a predicted and target sequence
using the Optimal String Alignment (OSA) method. The score ranges from 0 (no alignment)
to 1 (perfect alignment).
}
