% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing__data.R
\name{simplify_classes}
\alias{simplify_classes}
\title{Simplify Classes}
\usage{
simplify_classes(data, column, class_map)
}
\arguments{
\item{data}{A data frame. The data frame containing the column whose classes need to be simplified.}

\item{column}{A character string specifying the column name whose classes need to be simplified.}

\item{class_map}{A named character vector. A mapping of existing class names to their simplified versions.
The names of the vector represent the existing classes, and the values represent the
corresponding simplified classes.}
}
\value{
A modified version of the input data frame where the specified column's classes have been simplified
according to the provided class map.
}
\description{
Simplify the classes of a specific column in a data frame using a provided class map.
}
