% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing__data.R
\name{stratified_group_sampling_k_fold}
\alias{stratified_group_sampling_k_fold}
\title{Stratified Group Sampling}
\usage{
stratified_group_sampling_k_fold(df, kfolds = 10)
}
\arguments{
\item{df}{A data frame containing the dataset to be sampled.}
}
\value{
A list containing the indices of groups for training and testing.
}
\description{
Function to perform stratified group sampling
}
\section{Implementation is based on this kaggle kernel:}{
This function performs stratified group sampling on a dataset where each group has a target variable. It distributes the groups into folds while ensuring that the class proportions are maintained in each fold.
}

