% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_strata.R
\name{survey_strata}
\alias{survey_strata}
\title{Survey the layout and execution order of your project}
\usage{
survey_strata(project_path)
}
\arguments{
\item{project_path}{A path to strata project folder.}
}
\value{
dataframe housing the layout of your project based on the .tomls
}
\description{
\code{survey_strata()} will examine the \code{project_path} provided and
return a dataframe with the following information about the project:
\itemize{
\item \code{stratum_name}: the name of the stratum
\item \code{lamina_name}: the name of the lamina
\item \code{execution_order}: the order in which the stratum-lamina-code combination
will be executed
\item \code{script_name}: the name of the script to be executed
\item \code{script_path}: the path to the script
}
}
\examples{
\dontrun{
survey_strata("PATH/TO/PROJECT/FOLDER/")
}
}
