% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_toml_interaction.R
\name{survey_tomls}
\alias{survey_tomls}
\title{Find all toml files in a project}
\usage{
survey_tomls(project_path)
}
\arguments{
\item{project_path}{Path to the project}
}
\value{
an fs_path of all toml files
}
\description{
Find all toml files in a project
}
\examples{
\dontrun{
find_tomls("PATH/TO/PROJECT/FOLDER/")
}
}
