% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sequence__transition_match_F1.R
\name{tm_f1_score_custom_summary}
\alias{tm_f1_score_custom_summary}
\title{Calculate Transition Match F1 Score Summary}
\usage{
tm_f1_score_custom_summary(df)
}
\arguments{
\item{df}{A data frame containing the predicted values in the \code{prediction} column, the ground truth in the \code{target_variable} column,
and a grouping variable \code{nr} (e.g., borehole ID).}
}
\value{
A numeric value representing the mean Transition Match F1 score for all boreholes in the data frame (\code{df}).
}
\description{
This function calculates the mean Transition Match F1 score in the input data frame (\code{df}).
The score is computed by applying the \code{transition_match_f1} function to the \code{prediction} and \code{target_variable} columns for each borehole,
excluding rows where the target variable is "NN".
The final output is the mean F1 score for all boreholes in the data frame (\code{df}).
}
