% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction__neural_network.R
\name{train_model}
\alias{train_model}
\title{Train a Keras Model with Early Stopping}
\usage{
train_model(
  model,
  train_data,
  val_data,
  epochs = 100,
  batch_size = 16,
  patience = 10
)
}
\arguments{
\item{model}{A compiled Keras model to be trained.}

\item{train_data}{A list containing the training data. It should include \code{value_ragged_tensor} (input features)
and \code{target_ragged_tensor} (target labels).}

\item{val_data}{A list containing the validation data.}

\item{epochs}{An integer (default = 100) specifying the number of training epochs.}

\item{batch_size}{An integer (default = 16) specifying the batch size used during training.}

\item{patience}{An integer (default = 10) specifying the number of epochs with no improvement in validation loss before stopping the training.}
}
\value{
The history object containing the training process metrics (e.g., loss, accuracy).
}
\description{
This function trains a Keras model using the provided training and validation data. It includes early stopping to prevent
overfitting. The model training stops when the validation loss does not improve for a specified number of epochs (\code{patience}).
}
