% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sequence__transition_validation_score.R
\name{transition_validation_custom_summary}
\alias{transition_validation_custom_summary}
\title{Transition validation score Custom Summary}
\usage{
transition_validation_custom_summary(df, known_transitions)
}
\arguments{
\item{df}{A data frame or tibble containing columns \code{nr}, \code{prediction}}
}
\value{
A data frame with the mean perborehole transition validation score and the overall transition validation score.
}
\description{
This function calculates the overall and perborehole transition validation score.
}
