% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_extent__unit_match_F1.R
\name{um_f1_custom_summary}
\alias{um_f1_custom_summary}
\title{Compute Mean Unit Match F1 Score}
\usage{
um_f1_custom_summary(df)
}
\arguments{
\item{df}{A data frame containing \code{nr} (grouping variable), \code{target_variable} (true labels), and \code{prediction} (predicted labels).}
}
\value{
The mean Unit Match F1 score.
}
\description{
This function calculates the mean Unit Match F1 score for a dataset by grouping it based on \code{nr},
computing the Unit Match F1 score for each borehole "nr", and then averaging the results.
}
