% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_extent__UEVS.R
\name{unit_extent_validation_custom_summary}
\alias{unit_extent_validation_custom_summary}
\title{Calculate mean Unit Extent Validation Score per borehole}
\usage{
unit_extent_validation_custom_summary(df)
}
\arguments{
\item{df}{A data frame containing \code{prediction} (predicted labels).}
}
\value{
A numeric value representing the mean proportion of predicted units per borehole that match expected formations
}
\description{
This function serves as a wrapper for \code{unit_extent_validation}
}
