% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_extent__unit_match_F1.R
\name{unit_match_f1}
\alias{unit_match_f1}
\title{Compute Unit Match F1 Score}
\usage{
unit_match_f1(target_variable, prediction)
}
\arguments{
\item{target_variable}{A vector of true stratigraphic unit labels.}

\item{prediction}{A vector of predicted stratigraphic unit labels.}
}
\value{
The Unit Match F1 score, a value between 0 and 1.
}
\description{
This function calculates the Unit Match F1 score, which evaluates how well predicted stratigraphic units match the true stratigraphic units in a borehole.
It considers formations as True Positives (TP) if they appear in both the predicted and true sequences.
False Positives (FP) are formations predicted but not present in the ground truth, and False Negatives (FN) are formations in the ground truth but not predicted.
The metric ignores cases where \code{target_variable} is "NN".
}
