% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing__data.R
\name{units_in_coordinate}
\alias{units_in_coordinate}
\title{Identify Units at a Specific Coordinate}
\usage{
units_in_coordinate(units, x_28992, y_28992)
}
\arguments{
\item{units}{An \code{sf} object representing spatial layers, where each feature corresponds to a different unit. This object should have a \code{layer_name} column, which contains the names of the units.}

\item{y_28992}{A numeric value representing the Y-coordinate (northing) in the Amersfoort / RD New (EPSG: 28992) coordinate system.}

\item{28992}{A numeric value representing the X-coordinate (easting) in the Amersfoort / RD New (EPSG: 28992) coordinate system.}
}
\value{
A character vector containing the names of the units (from the \code{layer_name} column) that intersect with the provided coordinate. If no units intersect, the function returns an empty vector.
}
\description{
This function identifies the units that are present at a given coordinate within a spatial dataset. The function takes in a spatial object containing different units (e.g., geological layers) and checks which of these units intersect with the provided coordinate.
}
