% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_toml_interaction.R
\name{view_toml}
\alias{view_toml}
\title{View the contents of a toml file as a dataframe}
\usage{
view_toml(toml_path)
}
\arguments{
\item{toml_path}{Path to the toml file}
}
\value{
a dataframe of the toml file contents
}
\description{
View the contents of a toml file as a dataframe
}
\examples{
\dontrun{
view_toml("PATH/TO/TOML/FILE.toml")
}
}
