% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{build_lamina}
\alias{build_lamina}
\title{Add a lamina skeleton to your project space}
\usage{
build_lamina(lamina_name, stratum_path, order = 1, skip_if_fail = FALSE)
}
\arguments{
\item{lamina_name}{Name of your Lamina}

\item{stratum_path}{Path to the parent stratum}

\item{order}{Execution order inside of stratum}

\item{skip_if_fail}{Skip this lamina if it fails}
}
\value{
invisibly returns fs::path to lamina
}
\description{
Add a lamina skeleton to your project space
}
\examples{
\dontrun{
build_lamina("my_lamina_name", "PATH/TO/STRATUM/FOLDER/")
}
}
