% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{build_stratum}
\alias{build_stratum}
\title{Add a stratum to the project space}
\usage{
build_stratum(stratum_name, project_path, order = 1)
}
\arguments{
\item{stratum_name}{Name of stratum}

\item{project_path}{A path to strata project folder.}

\item{order}{Execution order, default is \code{1}}
}
\value{
invisibly returns fs::path to stratum
}
\description{
Add a stratum to the project space
}
\examples{
tmp <- fs::dir_create(fs::file_temp())
result <- build_stratum("my_stratum_name", tmp)
result
fs::dir_delete(tmp)
}
