% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-core.R
\name{log_error}
\alias{log_error}
\title{Wrapper around log_message for ERROR messages in the log}
\usage{
log_error(message)
}
\arguments{
\item{message}{A string containing a message to log.}
}
\value{
A message printed to stderr and an invisible character string copy of
the entire log error message.
}
\description{
\code{log_error()} does \emph{not} stop the execution of the script, but it does print
the message to stderr.
}
\examples{
log_error("This is an error message")
}
\seealso{
Other log: 
\code{\link{log_message}()},
\code{\link{log_total_time}()},
\code{\link{survey_log}()}
}
\concept{log}
