% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Entry point and automation target for your strata project}
\usage{
main(project_path, silent = FALSE)
}
\arguments{
\item{project_path}{A path to strata project folder}

\item{silent}{Suppress log messages? If \code{FALSE} (the default), log messages
will be printed to the console. If \code{TRUE}, log messages will be suppressed.}
}
\value{
invisible execution plan
}
\description{
Entry point and automation target for your strata project
}
\examples{
\dontrun{
main("/PATH/TO/PROJECT/FOLDER")
}
}
