% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Execute entire strata project}
\usage{
main(project_path, silent = FALSE)
}
\arguments{
\item{project_path}{A path to strata project folder.}

\item{silent}{Suppress log messages? If \code{FALSE} (the default), log messages
will be printed to the console. If \code{TRUE}, log messages will be suppressed.}
}
\value{
invisible execution plan.
}
\description{
\code{main()} will read the \code{.toml} files inside the \code{project_path} and begin
sourcing the \code{strata} and \code{laminae} in the order specified by the user,
with or without logging messages.

When a strata project is created \code{main.R} is added to the project root.
This script houses \code{main()}, and this file is the entry point to the project
and should be the target for automation. However, \code{main()} can be called
from anywhere, and users can opt to not use \code{main.R} at all.
}
\examples{
tmp <- fs::dir_create(fs::file_temp())
result <- strata::build_quick_strata_project(tmp, 1, 1)
main(tmp)
fs::dir_delete(tmp)
}
