% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Entry point and automation target for your strata project}
\usage{
main(project_path, silent = FALSE)
}
\arguments{
\item{project_path}{A path to automation project folder}

\item{silent}{A logical flag to suppress logging output}
}
\value{
invisible execution plan
}
\description{
Entry point and automation target for your strata project
}
\examples{
\dontrun{
main("/PATH/TO/PROJECT/FOLDER")
}
}
