% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata-package.R
\docType{package}
\name{strata-package}
\alias{strata}
\alias{strata-package}
\title{strata: Simple Data Pipelines For Simple Automation}
\description{
Tools for a simple, human-focused stratum framework for developers with only the most basic automation tool suite. The package is designed to get developers up and running quickly with the code they already have, and provide simple automation targets for use with tools like 'cron' or 'Task Scheduler'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asenetcky/strata}
  \item Report bugs at \url{https://github.com/asenetcky/strata/issues}
}

}
\author{
\strong{Maintainer}: Alexander Senetcky \email{asenetcky@gmail.com.com} (\href{https://orcid.org/0009-0009-3730-5397}{ORCID})

}
\keyword{internal}
