% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata-package.R
\docType{package}
\name{strata-package}
\alias{strata}
\alias{strata-package}
\title{strata: Simple Framework for Simple Automation}
\description{
Build a project framework for users with access to only the most basic of automation tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asenetcky/strata}
  \item \url{https://asenetcky.github.io/strata/}
  \item Report bugs at \url{https://github.com/asenetcky/strata/issues}
}

}
\author{
\strong{Maintainer}: Alexander Senetcky \email{asenetcky@gmail.com} (\href{https://orcid.org/0009-0009-3730-5397}{ORCID}) [copyright holder]

}
\keyword{internal}
