% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_strata.R
\name{survey_strata}
\alias{survey_strata}
\title{Survey the layout and execution order of your project}
\usage{
survey_strata(project_path)
}
\arguments{
\item{project_path}{path to your project folder}
}
\value{
data frame of the layout of your project based on the .tomls
}
\description{
Survey the layout and execution order of your project
}
\examples{
\dontrun{
survey_strata("PATH/TO/PROJECT/FOLDER/")
}
}
