% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_toml_interaction.R
\name{survey_tomls}
\alias{survey_tomls}
\title{Find all toml files in a project}
\usage{
survey_tomls(project_path)
}
\arguments{
\item{project_path}{A path to strata project folder.}
}
\value{
an fs_path object of all toml files.
}
\description{
Find all toml files in a project
}
\examples{
tmp <- fs::dir_create(fs::file_temp())
strata::build_quick_strata_project(tmp, 2, 3)
survey_tomls(tmp)
fs::dir_delete(tmp)
}
