% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_toml_interaction.R
\name{view_toml}
\alias{view_toml}
\title{View the contents of a toml file as a dataframe}
\usage{
view_toml(toml_path)
}
\arguments{
\item{toml_path}{Path to the toml file}
}
\value{
a dataframe of the toml file contents.
}
\description{
View the contents of a toml file as a dataframe
}
\examples{
tmp <- fs::dir_create(fs::file_temp())
strata::build_quick_strata_project(tmp, 2, 3)
proj_tomls <- survey_tomls(tmp)
purrr::map(proj_tomls, view_toml)
fs::dir_delete(tmp)
}
